requirements.txt
protobuf>=3.20,<4
langchain-huggingface>=0.0.3,<1
langchain>=0.2.17,<1
langchain_community>=0.2.19,<1
langchain-google-genai
python-dotenv
langchain_experimental>=0.0.65,<1
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime
flask
pysqlite3-wheels
chromadb
scikit-learn
numpy
langchain-core>=0.2.43,<1
langsmith>=0.1.112,<1
tenacity>=8.1.0,<9
